Fixes https://github.com/TheBusyBiscuit/Slimefun4/issues/375
